+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
+use App\Models\Profession;
 use App\Models\User;
 use Illuminate\Support\Facades\DB;
 
@@ -11,14 +12,15 @@ class UserController extends Controller
 {
     public function index()
     {
-        $users = User::ordered()
-            ->get();
+        $users = User::ordered()->get();
         return view('user.index', compact('users'));
     }
 
     public function create()
     {
-        return view('user.create');
+        $professions = Profession::ordered()->get();
+
+        return view('user.create', compact('professions'));
     }
 
     public function store(StoreUserRequest $request)
@@ -27,32 +29,40 @@ class UserController extends Controller
             $user = new User();
             $user->title = $request->title;
             $user->save();
+
+            $user->professions()->attach($request->professions);
         });
 
         return back();
     }
 
-    public function show($id)
+    public function show(User $user)
     {
-        $user = User::findOrFail($id);
-
         return view('user.show', compact('user'));
     }
 
-    public function update(UpdateUserRequest $request, $id)
+    public function edit(User $user)
     {
-        DB::transaction(function () use ($request, $id) {
-            $user = User::findOrFail($id);
-            $user->title = $request->title;
-            $user->save();
-        });
+        $professions = Profession::ordered()->get();
 
-        return back();
+        return view('user.edit', compact('user', 'professions'));
     }
 
-    public function destroy($id)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        User::findOrFail($id)->delete();
+        DB::transaction(function () use ($request, $user) {
+            $user->title = $request->title;
+            $user->save();
+
+            $user->professions()->sync($request->professions);
+        });
+
+        return redirect(route('users.show', $user->id));
+    }
+
+    public function destroy(User $user)
+    {
+        $user->delete();
         return redirect(route('users.index'));
     }
 }