+++ resources/views/components/selects/professions.blade.php
@@ -1,0 +1,9 @@
+@include('components.selects.template', [
+    'label' => 'Профессии',
+    'name' => 'professions',
+    'current' => $currentProfessions ?? [],
+    'elements' => $professions,
+    'id' => 'id',
+    'title' => 'title'
+])
+