+++ resources/views/genre/show.blade.php
@@ -74,25 +74,9 @@
         </div>
     </div>
 
-    <div class="modal fade" id="delete_genre" tabindex="-1" role="dialog" aria-labelledby="delete_genre_title"
-         aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="delete_genre_title">Modal title</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form action="{{route('genres.destroy', $genre->id)}}" method="POST">
-                        @method('DELETE')
-                        @csrf
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Отмена</button>
-                        <button type="submit" class="btn btn-danger">Удалить</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
+    @include('components.modals.delete_modal', [
+        'modalId' => 'delete_genre',
+        'modalTitle' => "Удалить жанр '$genre->title'",
+        'route' => route('genres.destroy', $genre->id)
+    ])
 @endsection