+++ resources/views/profession/show.blade.php
@@ -5,12 +5,11 @@
     @include('components.title_row', ['title' => $profession->title])
     <div class="row mb-3">
         <div class="col-sm-12">
-            <button type="button"
-                    class="btn btn-outline-primary btn-sm"
-                    data-toggle="modal"
-                    data-target="#edit_profession">
+            <a type="button"
+               class="btn btn-outline-primary btn-sm"
+               href="{{route('professions.edit', $profession->id)}}">
                 Редактировать
-            </button>
+            </a>
             @if(true)
                 <button type="button"
                         class="btn btn-outline-danger btn-sm"
@@ -23,63 +22,19 @@
 
     <div class="row">
         <div class="col-sm-12">
+            @include('components.users_table', ['users' => $profession->users])
+        </div>
+    </div>
+
+    <div class="row">
+        <div class="col-sm-12">
             @include('components.timestamps', ['element' => $profession])
         </div>
     </div>
 
-    <div class="modal fade" id="edit_profession" tabindex="-1" role="dialog" aria-labelledby="edit_profession_title"
-         aria-hidden="true">
-        <div class="modal-dialog modal-xl" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="edit_profession_title">Modal title</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form action="{{route('professions.update', $profession->id)}}" method="POST">
-                        @method('PATCH')
-                        @csrf
-                        <div class="form-group">
-                            <div class="row">
-                                <div class="col-sm-12">
-                                    <label for="title">Название</label>
-                                    <input type="text"
-                                           class="form-control"
-                                           name="title"
-                                           id="title"
-                                           value="{{$profession->title}}">
-                                </div>
-                            </div>
-                        </div>
-                        <button type="submit" class="btn btn-primary">Сохранить</button>
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Отмена</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="modal fade" id="delete_profession" tabindex="-1" role="dialog" aria-labelledby="delete_profession_title"
-         aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="delete_profession_title">Modal title</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form action="{{route('professions.destroy', $profession->id)}}" method="POST">
-                        @method('DELETE')
-                        @csrf
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Отмена</button>
-                        <button type="submit" class="btn btn-danger">Удалить</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
+    @include('components.modals.delete_modal', [
+        'modalId' => 'delete_profession',
+        'modalTitle' => "Удалить профессию '$profession->title'",
+        'route' => route('professions.destroy', $profession->id)
+    ])
 @endsection